Remove bugs for game continuation loading
This update addresses several issues that were preventing the proper loading of a game to be continued. Methods from the deleted DatabaseCallbacksOperations.py file have been moved to CallbackBase.py. In addition, relevant classes and files in the Callbacks directory have been updated to correctly use the centralized methods. A new feature to load details of a game is also added to the Callbacks directory. Necessary tests have been added as well.

+++ Repositories/GameRepository/PickleGamesRepository.py
@@ -38,7 +38,8 @@ class PickleGamesRepository(AbstractGamesRepository):
             games = self.list_games(suppress_logger = True)
             for game in games:
                 game_ts = dt.datetime.fromtimestamp(float(game))
-                if (dt.datetime.now() - game_ts).days > self.deleteAfterNDays:
+                # noinspection PyChainedComparisons
+                if self.deleteAfterNDays > 0 and (dt.datetime.now() - game_ts).days > self.deleteAfterNDays:
                     self.logger.info(f"Game '{game}' is older than {self.deleteAfterNDays} days")
                     self.delete_game(game)
             return func(self, *args, **kwargs)
